+++ src/main/java/boomcow/minezero/items/ArtifactFluteItem.java
@@ -31,10 +31,10 @@ public class ArtifactFluteItem extends Item {
         if (!level.isClientSide) {
             if (player instanceof ServerPlayer serverPlayer) {
 
-                boolean cooldownEnabled = level.getGameRules().getBoolean(ModGameRules.FLUTE_COOLDOWN_ENABLED);
+                boolean cooldownEnabled = level.getGameRules().getBoolean(ModGameRules.ARTIFACT_FLUTE_COOLDOWN_ENABLED);
                 if (cooldownEnabled) {
 
-                    int cooldownSeconds = level.getGameRules().getInt(ModGameRules.FLUTE_COOLDOWN_DURATION);
+                    int cooldownSeconds = level.getGameRules().getInt(ModGameRules.ARTIFACT_FLUTE_COOLDOWN_SECONDS);
                     int cooldownTicks = cooldownSeconds * 20;
 
                     if (serverPlayer.getCooldowns().isOnCooldown(this)) {